# core, trie, rpc: speed up tests